# swagger.yaml
openapi: 3.0.0
paths:
...
  "/v1/user/highlights/{id}/reactions":
    get:
      operationId: getAllHighlightUserReactions
      summary: Retrieves a highlight reactions for the authenticated user
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbUserHighlightReactionResponse"
        "400":
          description: Invalid request
        "404":
          description: Unable to get user highlight reactions
      tags: *a3
      security:
        - bearer: []
...
  /v1/repos/list:
    get:
      operationId: findAll
      summary: Finds all repos and paginates them
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
        - name: orderBy
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/RepoOrderFieldsEnum"
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbRepo"
      tags: *a8
...
  /v1/highlights/list:
    get:
      operationId: findAllHighlights
      summary: Finds all highlights and paginates them
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
        - name: repo
          required: false
          in: query
          description: Highlight Repo Filter
          example: open-sauced/insights
          schema:
            type: string
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbUserHighlight"
      tags:
        &a24
        - Highlights service
  /v1/highlights/featured:
    get:
      operationId: findAllFeaturedHighlights
      summary: Finds all featured highlights and paginates them
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
        - name: repo
          required: false
          in: query
          description: Highlight Repo Filter
          example: open-sauced/insights
          schema:
            type: string
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbUserHighlight"
      tags: *a24
  "/v1/highlights/{id}/featured":
    post:
      operationId: addAFeaturedHighlight
      summary: Add a highlight to the featured list
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbUserHighlight"
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid request
        "404":
          description: Unable to find highlight
      tags: *a24
      security:
        - bearer: []
    delete:
      operationId: removeAFeaturedHighlight
      summary: Remove a highlight from the featured list
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbUserHighlight"
        "400":
          description: Invalid request
        "404":
          description: Unable to find highlight
      tags: *a24
      security:
        - bearer: []
  /v1/highlights/repos/list:
    get:
      operationId: findAllHighlightRepos
      summary: Finds all highlight repos and paginates them
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbUserHighlightRepo"
      tags: *a24
  "/v1/highlights/{id}/reactions":
    get:
      operationId: getAllHighlightReactions
      summary: Retrieves total reactions for a highlight
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbUserHighlightReactionResponse"
        "400":
          description: Invalid request
        "404":
          description: Unable to get user highlight reactions
      tags: *a24
  /v1/highlights/top:
    get:
      operationId: findTopHighlights
      summary: Finds top highlights for the day range and paginates them
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: integer
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 1000
            default: 10
            type: integer
        - name: orderDirection
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/OrderDirectionEnum"
        - name: range
          required: false
          in: query
          description: Range in days
          schema:
            default: 30
            type: integer
      responses:
        "200":
          description: Successfully received model list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DbUserHighlight"
      tags: *a24
...
info:
  title: "@open-sauced/api.opensauced.pizza"
  description: >-
    
    ## Swagger-UI API Documentation


    This REST API can be used to create, read, update or delete data from the Open Sauced community platform.

    The Swagger-UI provides useful information to get started and an overview of all available resources.

    Each API route is clickable and has their own detailed description on how to use it.

    The base URL for the API is [api.opensauced.pizza](https://api.opensauced.pizza).


    [comment]: # (TODO: add bearer auth information)


    ## Rate limiting


    Every IP address is allowed to perform 5000 requests per hour.

    This is measured by saving the date of the initial request and counting all requests in the next hour.

    When an IP address goes over the limit, HTTP status code 429 is returned.

    The returned HTTP headers of any API request show the current rate limit status:


    header | description

    --- | ---

    `X-RateLimit-Limit` | The maximum number of requests allowed per hour

    `X-RateLimit-Remaining` | The number of requests remaining in the current rate limit window

    `X-RateLimit-Reset` | The date and time at which the current rate limit window resets in [UTC epoch seconds](https://en.wikipedia.org/wiki/Unix_time)


    [comment]: # (TODO: add pagination information)


    ## Common response codes


...

    The table below shows most common response codes you can receive from our endpoints.


...

...

...

...

...

...

...

...

...

...


...
servers:
...
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    PageMetaDto:
      type: object
      properties:
        page:
          type: integer
          description: The current page
          example: 1
        limit:
          type: integer
          description: The number of items per page
          example: 10
        itemCount:
          type: integer
          description: The number of items in the collection
          example: 100
        pageCount:
          type: integer
          description: The number of pages in the collection
          example: 10
        hasPreviousPage:
          type: boolean
          description: Flag indicating if there is a previous page
          example: false
        hasNextPage:
          type: boolean
          description: Flag indicating if there is a next page
          example: true
      required:
        - page
        - limit
        - itemCount
        - pageCount
        - hasPreviousPage
        - hasNextPage
    PageDto:
      type: object
      properties:
        data:
          type: array
          items:
            type: array
        meta:
          $ref: "#/components/schemas/PageMetaDto"
      required:
        - data
        - meta
...
    OrderDirectionEnum:
      type: string
      enum:
        - ASC
        - DESC
    DbUserHighlight:
      type: object
      properties:
        id:
          type: integer
          description: User Highlight identifier
          example: 237133
        user_id:
          type: integer
          description: User identifier
          example: 237133
        url:
          type: string
          description: Highlight Pull Request URL
          example: github.com/open-sauced/insights/pull/1
        title:
          type: string
          description: Highlight Title
          example: My First PR!
        highlight:
          type: string
          description: Highlight Text
          example: |-2
            
                I made my first open source pull request!

                github.com/open-sauced/insights/pull/1
        type:
          type: string
          description: Highlight type
          example: pull_request
        pinned:
          type: boolean
          description: Whether the highlight is pinned to the top
          example: false
        featured:
          type: boolean
          description: Whether the highlight is featured or not
          example: false
        created_at:
          format: date-time
          type: string
          description: Timestamp representing highlight creation date
          example: 2023-01-19 13:24:51.000000
        updated_at:
          format: date-time
          type: string
          description: Timestamp representing highlight updated date
          example: 2023-01-19 13:24:51.000000
        deleted_at:
          format: date-time
          type: string
          description: Timestamp representing highlight deletion date
          example: 2023-01-19 13:24:51.000000
        shipped_at:
          format: date-time
          type: string
          description: Timestamp representing highlight shipped date
          example: 2023-01-19 13:24:51.000000
        full_name:
          type: string
          description: Highlight Repo Full Name
          example: open-sauced/insights
        name:
          type: string
          description: Highlight User Full Name
          example: Brian Douglas
        login:
          type: string
          description: Highlight User Login
          example: bdougie
        tagged_repos:
          description: An array of full-names of tagged repositories
          example:
            - open-sauced/insights
            - open-sauced/ai
          type: array
          items:
            type: string
      required:
        - id
        - user_id
        - highlight
        - type
        - tagged_repos
    DbRepo:
      type: object
      properties:
        id:
          type: integer
          description: Repository identifier
          example: 71359796
        user_id:
          type: integer
          description: Owner user identifier
          example: 57568598
        size:
          type: integer
          description: Total size in bytes
          example: 274322
        issues:
          type: integer
          description: Total number of issues
          example: 274
        stars:
          type: integer
          description: Total number of stars
          example: 5
        forks:
          type: integer
          description: Total number of forks
          example: 1
        watchers:
          type: integer
          description: Total number of watchers
          example: 5473
        subscribers:
          type: integer
          description: Total number of subscribers
          example: 11756
        network:
          type: integer
          description: Total number of network usages
          example: 4
        is_fork:
          type: boolean
          description: Flag indicating repo is a fork
          example: false
        is_private:
          type: boolean
          description: Flag indicating repo is private
          example: false
        is_template:
          type: boolean
          description: Flag indicating repo is a template
          example: false
        is_archived:
          type: boolean
          description: Flag indicating repo is archived
          example: false
        is_disabled:
          type: boolean
          description: Flag indicating repo is disabled
          example: false
        has_issues:
          type: boolean
          description: Flag indicating repo has issues enabled
          example: false
        has_projects:
          type: boolean
          description: Flag indicating repo has projects enabled
          example: false
        has_downloads:
          type: boolean
          description: Flag indicating repo has downloads enabled
          example: false
        has_wiki:
          type: boolean
          description: Flag indicating repo has wiki enabled
          example: false
        has_pages:
          type: boolean
          description: Flag indicating repo has pages enabled
          example: false
        has_discussions:
          type: boolean
          description: Flag indicating repo has discussions enabled
          example: false
        created_at:
          format: date-time
          type: string
          description: Timestamp representing repository creation
          example: 2016-10-19 13:24:51.000000
        updated_at:
          format: date-time
          type: string
          description: Timestamp representing repository last update
          example: 2022-08-28 22:04:29.000000
        pushed_at:
          format: date-time
          type: string
          description: Timestamp representing repository last push
          example: 2022-08-28 22:04:39.000000
        default_branch:
          type: string
          description: Repository default branch
          example: main
        node_id:
          type: string
          description: Repository GQL node id
          example: MDEwOlJlcG9zaXRvcnk3MTM1OTc5Ng==
        git_url:
          type: string
          description: Repository git url
          example: git://github.com/open-sauced/insights.git
        ssh_url:
          type: string
          description: Repository ssh url
          example: git@github.com:open-sauced/insights.git
        clone_url:
          type: string
          description: Repository clone url
          example: https://github.com/open-sauced/insights.git
        svn_url:
          type: string
          description: Repository svn url
          example: https://github.com/open-sauced/insights
        mirror_url:
          type: string
          description: Repository mirror url
          example: null
        name:
          type: string
          description: Repository unique name
          example: open-sauced
        full_name:
          type: string
          description: Repository full name
          example: open-sauced/insights
        description:
          type: string
          description: Repository short description
          example: 🍕This is a project to identify your next open source contribution! 🍕
        language:
          type: string
          description: Repository programming language
          example: JavaScript
        license:
          type: string
          description: Repository SPDX license
          example: MIT
        url:
          type: string
          description: Repository GitHub linked URL
          example: https://api.github.com/repos/open-sauced/insights
        homepage:
          type: string
          description: Repository GitHub homepage
          example: https://app.opensauced.pizza
        topics:
          description: Repository GitHub topics
          example:
            - open-sauced
            - open-source
            - github
          type: array
          items:
            type: string
        open_prs_count:
          type: integer
          description: Repository number of open PRs
          example: 5
        closed_prs_count:
          type: integer
          description: Repository number of closed PRs
          example: 173
        merged_prs_count:
          type: integer
          description: Repository number of merged PRs
          example: 276
        draft_prs_count:
          type: integer
          description: Repository number of draft PRs
          example: 2
        spam_prs_count:
          type: integer
          description: Repository number of spam PRs
          example: 2
        pr_velocity_count:
          type: integer
          description: Repository average open/close time for PRs
          example: 0
        pr_active_count:
          type: integer
          description: Number of non-closed PRs updated within the day range
          example: 0
      required:
        - id
        - user_id
        - size
        - issues
        - stars
        - forks
        - watchers
        - subscribers
        - network
        - is_fork
        - is_private
        - is_template
        - is_archived
        - is_disabled
        - has_issues
        - has_projects
        - has_downloads
        - has_wiki
        - has_pages
        - has_discussions
        - default_branch
        - node_id
        - git_url
        - ssh_url
        - clone_url
        - svn_url
        - name
        - full_name
        - description
        - language
        - license
        - url
        - homepage
        - topics
...
    DbUserHighlightReactionResponse:
      type: object
      properties:
        emoji_id:
          type: string
          description: Emoji identifier
          example: uuid-v4
        reaction_count:
          type: integer
        reaction_users:
          type: array
          items:
            required: false
            description: Usernames of users who reacted with this emoji
            example:
              - RitaDee
              - diivi
            type: array
            items:
              type: string
      required:
        - emoji_id
        - reaction_users
...
    DbUserHighlightRepo:
      type: object
      properties:
        full_name:
          type: string
          description: Highlight Repo Full Name
          example: open-sauced/insights
...
    RepoOrderFieldsEnum:
      type: string
      enum:
        - issues
        - stars
        - watchers
        - subscribers
        - created_at
        - updated_at
        - name
        - contributionsCount
        - votesCount
        - submissionsCount
        - stargazersCount
        - starsCount
...